+++ kafka-connect/kafka-connect-transforms/src/main/java/org/debezium/connector/mongodb/transforms/ArrayEncoding.java
@@ -1,0 +1,70 @@
+/*
+ *
+ *  * Licensed to the Apache Software Foundation (ASF) under one
+ *  * or more contributor license agreements.  See the NOTICE file
+ *  * distributed with this work for additional information
+ *  * regarding copyright ownership.  The ASF licenses this file
+ *  * to you under the Apache License, Version 2.0 (the
+ *  * "License"); you may not use this file except in compliance
+ *  * with the License.  You may obtain a copy of the License at
+ *  *
+ *  *   http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing,
+ *  * software distributed under the License is distributed on an
+ *  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ *  * KIND, either express or implied.  See the License for the
+ *  * specific language governing permissions and limitations
+ *  * under the License.
+ *
+ */
+package org.debezium.connector.mongodb.transforms;
+
+public enum ArrayEncoding {
+  ARRAY("array"),
+  DOCUMENT("document");
+
+  private final String value;
+
+  ArrayEncoding(String value) {
+    this.value = value;
+  }
+
+  public String getValue() {
+    return value;
+  }
+
+  /**
+   * Determine if the supplied value is one of the predefined options.
+   *
+   * @param value the configuration property value; may not be null
+   * @return the matching option, or null if no match is found
+   */
+  public static ArrayEncoding parse(String value) {
+    if (value == null) {
+      return null;
+    }
+    value = value.trim();
+    for (ArrayEncoding option : ArrayEncoding.values()) {
+      if (option.getValue().equalsIgnoreCase(value)) {
+        return option;
+      }
+    }
+    return null;
+  }
+
+  /**
+   * Determine if the supplied value is one of the predefined options.
+   *
+   * @param value the configuration property value; may not be null
+   * @param defaultValue the default value; may be null
+   * @return the matching option, or null if no match is found and the non-null default is invalid
+   */
+  public static ArrayEncoding parse(String value, String defaultValue) {
+    ArrayEncoding mode = parse(value);
+    if (mode == null && defaultValue != null) {
+      mode = parse(defaultValue);
+    }
+    return mode;
+  }
+}
